combine watch and audit into one component
This means we can't heal signatures that are lost, but removes the need for a queue and multiple processes.

TODO:  we need new namespace statuses to indicate that we lost a signature, and so cannot continue auditing, but we didn't detect anything wrong with the directory. The intent is that the directory must be re-audited if this happens